Changed directory, created the db file
